Let TableView deduce type of result set column
Set the data type held in columns to Object instead of String. This way numeric columns will be stored as actual numbers, and so sorting will work correctly on these columns.

Added the compressed imdb.db file, changed data types of movies.us_gross and movies.worldwide_gross, because some movies are into the high 2billions, they exceed the max value of Integers, and are causing some fucky sorting problems in the TableView. Set them to Real instead of Integer.