imdb.db changed
broke children/spouses/relatives/parents into 2 relationships each.

We will end up with tons of duplicated records in the person table, since for example we have no way of knowing if a certain non-famous child "John" is different or the same as any other "Jonh"s.

Assume:
| Parent            | Child  |
|-------------------|--------|
| Will Smith        | Johnny |
| Jada Pinket Smith | Johnny |

How would we know if "Johnny" is the same person, or a different person?