Changed QueryController to abstract class
When we want to do a query, we will call the execute method of the QueryController superclass, which will call the CreateQuery method, which simply returns a string.

The QueryController will then run the actual query, and handle the logging.

We can create another util class if we ever actually end up needing true PreparedStatements.